+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Changed
+- Detect name collisions #1
 
 ## [0.0.1] - 2024-04-18
 ### Added